refactor(web): split pipeline.py into four focused route modules
web/routes/pipeline.py (2202 lines) carried the dashboard, Decisions tab
(constants/simulate), beets-distance, and resolve-rg/replace handlers
alongside core pipeline CRUD. #522 splits each concern into its own module
(pipeline_dashboard.py, decisions.py, beets_distance.py,
release_identity_routes.py) mirroring the existing web/js/*.js split,
wires them into server.py's merge_registries, and moves/renames the
matching contract tests 1:1 (test count unchanged: 88 before, 88 after).

No behavior change — pure move, plus doc-comment path fixes in
lib/replace_status.py, lib/quality/decision_tree.py, web/routes/youtube.py,
and scripts/pipeline_cli/youtube.py that referenced the old file location.
